Rename rootReducer.js to be index.js
